+++ CHANGELOG.md
@@ -41,6 +41,10 @@ Format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and ver
 
 ## [Unreleased]
 
+### Added
+
+- Ctrl/Cmd+R sends the open REST request, the same as Ctrl/Cmd+Enter.
+
 ## [0.0.14] - 2026-08-22
 
 ### Added